feat: hooks로 비지니스 로직 분리 및 카테고리 기능 적용
