+++ src/pages/FindProjectPage.tsx
@@ -1,26 +1,18 @@
-import { useEffect, useState } from 'react';
-import { firebaseGetProjectDataRequest } from 'apis/boardService';
+import { useFindProject } from 'hooks';
 import WebContainer from 'components/common/WebContainer';
 import FindProjectHeader from 'components/findproject/FindProjectHeader';
 import FindProjectList from 'components/findproject/FindProjectList';
 import styled from '@emotion/styled';
 
 const FindProjectPage = () => {
-  const [projects, setProjects] = useState([]);
-  const [category, setCategory] = useState('기획');
-  const [toggle, setToggle] = useState(false);
-
-  useEffect(() => {
-    firebaseGetProjectDataRequest(setProjects);
-  }, []);
-
-  const handleCategoryClick = (e: any) => {
-    setCategory(e.target.name);
-  };
-
-  const handleToggleClick = () => {
-    setToggle((prev) => !prev);
-  };
+  const {
+    filterProjects,
+    category,
+    toggle,
+    handleCategoryClick,
+    handleToggleClick,
+    handleNavigateToProjectDetail,
+  } = useFindProject();
 
   return (
     <FindProjectPageWrapper>
@@ -31,7 +23,11 @@ const FindProjectPage = () => {
           onCategoryClickEvent={handleCategoryClick}
           onToggleClickEvent={handleToggleClick}
         />
-        <FindProjectList projects={projects} />
+        <FindProjectList
+          projects={filterProjects}
+          toggle={toggle}
+          onNavigateToProjectDetailEvent={handleNavigateToProjectDetail}
+        />
       </WebContainer>
     </FindProjectPageWrapper>
   );
